chore: reform the folder structure
