+++ packages/react-styled-core/src/shared/styled-system/config.js
